Add BGG identity validation resource
This adds a backend REST api to attempt to associate a BGG user name with a user account.
We generate a one time password that they must add to their BGG profile somewhere so we can
verify their identity.

It also introduces rabbitmq messaging between web and worker with the worker responsible for
making api calls to BGG.